First attempt at separating out AWS F1 bitstream building
