+++ deploy/firesim
@@ -99,9 +99,16 @@ def buildafi(globalbuildconf):
     """ Starting from local Chisel, build an AFI for all of the specified
     hardware configs. """
 
-    for buildconf in globalbuildconf.get_builds_list():
-        execute(replace_rtl, buildconf, hosts=['localhost'])
-        execute(build_driver, buildconf, hosts=['localhost'])
+    # TODO: AJG:
+    # Problem... how to build things on localhost in parallel?
+    # currently there is a mapping from ip -> buildconfig
+    # this breaks down if multiple things are building on localhost...
+    #for buildconf in globalbuildconf.get_builds_list():
+    #    execute(replace_rtl, buildconf, hosts=['localhost'])
+    #    execute(build_driver, buildconf, hosts=['localhost'])
+    num_builds = len(globalbuildconf.get_builds_list())
+    execute(globalbuildconf.replace_rtl, hosts=['localhost'] * num_builds)
+    execute(globalbuildconf.build_driver, hosts=['localhost'] * num_builds)
 
     @parallel
     def instance_liveness():
@@ -138,7 +145,7 @@ def buildafi(globalbuildconf):
     execute(instance_liveness, hosts=globalbuildconf.get_build_instance_ips())
 
     # run builds, then terminate instances
-    execute(aws_build, globalbuildconf, hosts=globalbuildconf.get_build_instance_ips())
+    execute(globalbuildconf.build_bitstream, hosts=globalbuildconf.get_build_instance_ips())
 
 
 def tar2afi(globalbuildconf):
