+++ deploy/sample-backup-configs/sample_config_build_recipes.ini
@@ -8,12 +8,17 @@
 # NB: This has a faster host-clock frequency than the NIC-based design, because
 # its uncore runs at half rate relative to the tile.
 [f1-firesim-rocket-quadcore-no-nic-l2-llc4mb-ddr3-1]
+# DESIGN SPECIFIC
 DESIGN=FireSim
 TARGET_CONFIG=DDR3FRFCFSLLC4MB_WithDefaultFireSimBridges_WithFireSimTestChipConfigTweaks_chipyard.QuadRocketConfig
 deploytriplet=None
+# RUN PLATFORM SPECIFIC
 PLATFORM_CONFIG=WithAutoILA_F40MHz_BaseF1Config
 postbuildhook=
 s3bucketname=firesim-local
+fpgaplatform=F1BitBuilder
+# BUILD PLATFORM SPECIFIC
+buildhost=localbuildhost
 
 [f1-firesim-rocket-quadcore-no-nic-l2-llc4mb-ddr3-2]
 # DESIGN SPECIFIC
@@ -24,6 +29,7 @@ deploytriplet=None
 PLATFORM_CONFIG=WithAutoILA_F35MHz_BaseF1Config
 postbuildhook=
 s3bucketname=firesim-local
+fpgaplatform=F1BitBuilder
 # BUILD PLATFORM SPECIFIC
 buildhost=localbuildhost
 
@@ -34,4 +40,5 @@ deploytriplet=None
 PLATFORM_CONFIG=WithAutoILA_F30MHz_BaseF1Config
 postbuildhook=
 s3bucketname=firesim-local
+fpgaplatform=F1BitBuilder
 buildhost=localbuildhost
